Refactor database.py.
