+++ src/_pytask/database.py
@@ -6,12 +6,10 @@ from pathlib import Path
 from typing import Any
 
 import click
-import networkx as nx
 from _pytask.click import EnumChoice
 from _pytask.config import hookimpl
-from _pytask.dag import node_and_neighbors
+from _pytask.database_utils import create_database
 from click import Context
-from pony import orm
 
 
 class _DatabaseProviders(enum.Enum):
@@ -20,48 +18,6 @@ class _DatabaseProviders(enum.Enum):
     MYSQL = "mysql"
     ORACLE = "oracle"
     COCKROACH = "cockroach"
-
-
-db = orm.Database()
-
-
-class State(db.Entity):  # type: ignore
-    """Represent the state of a node in relation to a task."""
-
-    task = orm.Required(str)
-    node = orm.Required(str)
-    state = orm.Required(str)
-
-    orm.PrimaryKey(task, node)
-
-
-def create_database(
-    provider: str, filename: str, *, create_db: bool, create_tables: bool
-) -> None:
-    """Create the database.
-
-    Raises
-    ------
-    orm.BindingError
-        Raise if database is already bound.
-
-    """
-    try:
-        db.bind(provider=provider, filename=filename, create_db=create_db)
-        db.generate_mapping(create_tables=create_tables)
-    except orm.BindingError:
-        pass
-
-
-@orm.db_session
-def create_or_update_state(first_key: str, second_key: str, state: str) -> None:
-    """Create or update a state."""
-    try:
-        state_in_db = State[first_key, second_key]  # type: ignore
-    except orm.ObjectNotFound:
-        State(task=first_key, node=second_key, state=state)
-    else:
-        state_in_db.state = state
 
 
 def _database_filename_callback(
@@ -128,10 +84,3 @@ def pytask_parse_config(config: dict[str, Any]) -> None:
 def pytask_post_parse(config: dict[str, Any]) -> None:
     """Post-parse the configuration."""
     create_database(**config["database"])
-
-
-def update_states_in_database(dag: nx.DiGraph, task_name: str) -> None:
-    """Update the state for each node of a task in the database."""
-    for name in node_and_neighbors(dag, task_name):
-        node = dag.nodes[name].get("task") or dag.nodes[name]["node"]
-        create_or_update_state(task_name, node.name, node.state())
