+++ tests/test_profile.py
@@ -6,7 +6,7 @@ from pathlib import Path
 
 import pytest
 from _pytask.cli import cli
-from _pytask.database import create_database
+from _pytask.database_utils import create_database
 from _pytask.profile import _to_human_readable_size
 from _pytask.profile import Runtime
 from pony import orm
